# backend/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ChGK-Helper API documentation",
        "version": "0.1.0",
        "description": "https://github.com/Minuta18/ChGK-helper",
        "contact": {
            "email": "ig.samsonov10@yandex.ru"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/Minuta18/ChGK-helper/blob/main/LICENSE"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:5000/api/v1"
        }
    ],
...
    "paths": {
...
        "/questions": {
            "get": {
                "tags": [
                    "questions"
                ],
                "summary": "Gets multiple questions by a given page",
                "description": "Gets multiple questions from a specified page. A page is questions from n + 1 to n + page_size.",
                "operationId": "get_question_by_page",
                "parameters": [
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Count of questions in one page",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/health_check"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid arguments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions/{question_id}": {
            "get": {
                "tags": [
                    "questions"
                ],
                "summary": "Gets question by given id",
                "description": "Returns question by given id (int). If question not found returns 404 error.",
                "operationId": "get_question_by_id",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "description": "ID of question to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/get_question_by_id"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "questions"
                ],
                "summary": "Edits question by given id",
                "description": "Edits question by given id",
                "operationId": "changeQuestionById",             
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "description": "Id of question to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "New text",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Было 2 козла. Сколько?"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "New comment",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Автор - стеапин"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/get_question_by_id"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid arguments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "questions"
                ],
                "summary": "Deletes question by given id",
                "description": "Deletes question by given id",
                "operationId": "deleteQuestionById",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "description": "Id of question to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/change_password"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "health_check": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "working",
                            "not working"
                        ],
                        "example": "working"
                    }   
                }
            },
...
            "error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "detail": {
                        "type": "string",
                        "example": "Error description"
                    }
                }
            },
...
            "change_password": {
                "type": "object",
                "properties": {
                    "old_password": {
                        "type": "string",
                        "example": "12345"
                    },
                    "new_password": {
                        "type": "string",
                        "example": "12345678"
                    }
                }
            },
...
            "question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Было 2 козла. Сколько?"
                    },
                    "comment": {
                        "type": "string",
                        "example": "Автор - стеапин"
                    }
                }
            },
            "get_question_by_id": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "question": {
                        "type": "object",
                        "$ref": "#/components/schemas/question"
                    }
                }
            },
...
        }
    }
}